Trying to Fix Nav Bar Beta
Using div made the main text jump to the bottom of the page. Planning
to ignore this because it’s late at night. One thing I wanted to do was
to check if the class needed to include the list items (li). Hopefully
this makes the text display inline and not bullet-pointed.